+++ src/components/center/ticket/Form/TicketFormComponent.tsx
@@ -4,6 +4,8 @@ import { useNavigate } from 'react-router-dom';
 
 import { LessonTypeEnum, TermUnitEnum, Ticket_put_body, Tickets_request, tickets_create } from '@apis/ticketsAPIs';
 import { Button } from '@components/common/Button';
+import { InputField, Unit } from '@components/common/InputField';
+import { SelectField } from '@components/common/SelectField';
 import useInput from '@hooks/utils/useInput';
 import { ValidationProps, useValidation } from '@hooks/utils/useValidation';
 import {
@@ -14,9 +16,6 @@ import {
   LabelNotice,
 } from '@styles/center/ticketFormStyle';
 import { FormContentWrap, SC, TopTitleWrap } from '@styles/styles';
-
-import { InputField, Unit } from './InputField';
-import { SelectField } from './SelectField';
 
 export type TicketFormDataType = Tickets_request | Ticket_put_body;
